Extract CardLink component to own file
